conversation API backend update
